requirements.txt
# VPS Deployment Requirements for AI Trading Bot
# Optimized for 16GB RAM / 160GB NVMe storage

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# ML frameworks
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Resource management
psutil>=5.8.0
schedule>=1.1.0
joblib>=1.1.0

# Data processing
pyarrow>=6.0.0
feather-format>=0.4.1

# Monitoring and logging
structlog>=21.5.0
python-json-logger>=2.0.0

# API and networking
requests>=2.26.0
websockets>=10.0
aiohttp>=3.8.0

# Utilities
python-dotenv>=0.19.0
click>=8.0.0
colorama>=0.4.4

# Web framework and API (Flask ecosystem)
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
Flask-Assets==2.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.2
Flask-Limiter==3.5.0
Flask-Mail==0.9.1

# Database and async operations
psycopg2-binary==2.9.9
asyncpg>=0.28.0
redis>=4.5.0
python-socketio==5.8.0
eventlet==0.33.3

# Security and utilities
cryptography>=41.0.0
itsdangerous==2.1.2
cachetools==5.4.0
ujson==5.11.0
cssmin==0.2.0
jsmin==3.0.1

# Trading and market data
python-binance==1.0.17

# Visualization and monitoring
matplotlib==3.8.2
seaborn==0.13.0
prometheus-client==0.19.0

# Memory profiling and optimization
memory-profiler>=0.61.0
line-profiler>=4.1.0
pympler>=1.0.1
autoflake>=2.2.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
mypy>=0.930
ruff==0.1.5
pytest-cov==4.1.0
pre-commit==3.6.0

# Advanced scheduling
APScheduler>=3.10.0
tenacity>=8.2.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.65.0
watchdog>=3.0.0

# Quality Diversity Optimization
ribs[all]>=0.5.0

# Visualization for RIBS
plotly>=5.17.0
dash>=2.14.0

# Parallel processing
concurrent-executor>=0.2.0

# Optional: GPU acceleration (uncomment if using neural emitters)
# torch>=2.0.0
